Refactor and add exit code expansion to heredoc
