+++ srcs/executor/builtins/builtin_echo.c
